requirements.txt
Flask
flask_mail
requests
shapely
pillow
piexif
plotly
virtualenv
flask
pymongo
passlib
tailwindcss
flask_wtf
wtforms
flask_mail
jsonify
google-api-python-client
google-auth
oauth2client
opencv-python
gpx_converter
pandas
numpy
Pillow
watchdog
gpxpy
Flask-Session
torch
ultralytics
pymediainfo
exif
ffmpeg-python
waitress
python-dotenv
flask_limiter
env